ndctl: enable 'travis-ci'
Add a .travis.yml config to enable Travis CI.
Also add the Travis CI  status 'badge' to README.md

For now, this only does a compile test. It can be built upon in the
future to do more runtime testing. Since the travis environment uses
Ubuntu, this is useful regardless, as it will immediately catch any
build failures in an Ubuntu environment.

Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>